Leverage the fact that Integer.MAX_VALUE is prime
To optimize away the bounds check in the nextPrime() loop. Since
Integer.MAX_VALUE is a prime, nextPrime() will always return
Integer.MAX_VALUE rather than overflow.

This is a tight loop, so that optimization actually saves 20 or 30
milliseconds. I know that doesn't sound like much, but we've already
gotten most of the low-hangning fruit at this point :)